backend/requirements.txt
Flask==3.0.0
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
redis==5.0.1
PyMySQL==1.1.0
python-dotenv==1.0.0
bcrypt==4.1.2
PyJWT==2.8.0 